feat: upload course image in manager course, logic button edit, delete Assignment

+++ front_end/fe-ae-lecture-student/hooks/assignment/useDeleteAssignment.ts
@@ -1,15 +1,22 @@
 "use client";
 
+import { AssignmentService } from "@/services/assignment.services";
+import { AssignmentStatus, DeleteAssignmentResponse } from "@/types/assignments/assignment.response";
 import { useState } from "react";
 import { toast } from "sonner";
-import { AssignmentService } from "@/services/assignment.services";
-import { DeleteAssignmentResponse } from "@/types/assignments/assignment.response";
 
 export function useDeleteAssignment() {
   const [loading, setLoading] = useState(false);
 
-  const deleteAssignment = async (id: string): Promise<DeleteAssignmentResponse | null> => {
+  const deleteAssignment = async (
+    id: string,
+    status?: AssignmentStatus
+  ): Promise<DeleteAssignmentResponse | null> => {
     if (loading) return null;
+    if (status && status !== AssignmentStatus.Draft) {
+      toast.error("Only Draft assignments can be deleted.");
+      return null;
+    }
     setLoading(true);
     try {
       const res = await AssignmentService.remove(id);
